+++ webapp/src/App.tsx
@@ -51,8 +51,6 @@ import UnclaimedOrderList from "./pages/UnclaimedOrderList";
 import Utilities from "./pages/Utilities";
 import applyHocs from "./shared/applyHocs";
 import { installPromiseExtras } from "./shared/bluejay";
-import Redirect from "./shared/react/Redirect";
-import renderComponent from "./shared/react/renderComponent";
 import BackendGlobalsProvider from "./state/BackendGlobalsProvider";
 import GlobalViewStateProvider from "./state/GlobalViewStateProvider";
 import OfferingProvider from "./state/OfferingProvider";
@@ -62,6 +60,8 @@ import withMetatags from "./state/withMetatags";
 import withPageLayout from "./state/withPageLayout";
 import withProps from "./state/withProps";
 import withScreenLoaderMount from "./state/withScreenLoaderMount";
+import Redirect from "./uir/Redirect";
+import renderComponent from "./uir/renderComponent";
 import React from "react";
 import { HelmetProvider } from "react-helmet-async";
 import { unstable_HistoryRouter as Router, Routes, Route } from "react-router-dom";
