+++ webapp/src/components/OrderDetail.tsx
@@ -5,8 +5,6 @@ import LayoutContainer from "../components/LayoutContainer";
 import SumaImage from "../components/SumaImage";
 import { dt, t } from "../localization";
 import { dayjs } from "../modules/dayConfig";
-import ScrollTopOnMount from "../shared/ScrollToTopOnMount";
-import Money from "../shared/react/Money";
 import useErrorToast from "../state/useErrorToast";
 import useScreenLoader from "../state/useScreenLoader";
 import useToggle from "../state/useToggle";
@@ -18,6 +16,8 @@ import Form from "../ui/Form";
 import FormCheck from "../ui/FormCheck";
 import FormGroup from "../ui/FormGroup";
 import Stack from "../ui/Stack";
+import Money from "../uir/Money";
+import ScrollTopOnMount from "../uir/ScrollToTopOnMount";
 import PressAndHold from "./PressAndHold";
 import isEmpty from "lodash/isEmpty";
 import React from "react";
