+++ webapp/src/hocs/authRedirects.tsx
@@ -1,6 +1,6 @@
-import Redirect from "../shared/react/Redirect";
 import useLoginRedirectLink from "../state/useLoginRedirectLink";
 import useUser from "../state/useUser";
+import Redirect from "../uir/Redirect";
 import React from "react";
 import { useLocation } from "react-router-dom";
