+++ webapp/src/pages/FoodCheckout.tsx
@@ -11,9 +11,7 @@ import PageLoader from "../components/PageLoader";
 import SumaImage from "../components/SumaImage";
 import { dt, t } from "../localization";
 import idempotency from "../modules/idempotency";
-import ScrollTopOnMount from "../shared/ScrollToTopOnMount";
 import { anyMoney } from "../shared/money";
-import Money from "../shared/react/Money";
 import useAsyncFetch from "../state/useAsyncFetch";
 import useBackendGlobals from "../state/useBackendGlobals";
 import useErrorToast from "../state/useErrorToast";
@@ -26,6 +24,8 @@ import Form from "../ui/Form";
 import FormControlFeedback from "../ui/FormControlFeedback";
 import FormGroup from "../ui/FormGroup";
 import Stack from "../ui/Stack";
+import Money from "../uir/Money";
+import ScrollTopOnMount from "../uir/ScrollToTopOnMount";
 import clsx from "clsx";
 import find from "lodash/find";
 import isEmpty from "lodash/isEmpty";
