+++ webapp/src/pages/FoodDetails.tsx
@@ -11,10 +11,10 @@ import SumaImage from "../components/SumaImage";
 import { dt, t } from "../localization";
 import makeTitle from "../modules/makeTitle";
 import { anyMoney, intToMoney } from "../shared/money";
-import Money from "../shared/react/Money";
 import useOffering from "../state/useOffering";
 import Col from "../ui/Col";
 import Row from "../ui/Row";
+import Money from "../uir/Money";
 import clsx from "clsx";
 import find from "lodash/find";
 import React from "react";
