+++ webapp/src/pages/LedgersOverview.tsx
@@ -6,7 +6,6 @@ import PageHeading from "../components/PageHeading";
 import PageLoader from "../components/PageLoader";
 import LedgerItemModal from "../components/ledger/LedgerItemModal";
 import { t } from "../localization";
-import Money from "../shared/react/Money";
 import useAsyncFetch from "../state/useAsyncFetch";
 import useHashSelector from "../state/useHashSelector";
 import useListQueryControls from "../state/useListQueryControls";
@@ -16,6 +15,7 @@ import DropdownMenu from "../ui/DropdownMenu";
 import DropdownToggle from "../ui/DropdownToggle";
 import Stack from "../ui/Stack";
 import Table from "../ui/Table";
+import Money from "../uir/Money";
 import clsx from "clsx";
 import dayjs from "dayjs";
 import find from "lodash/find";
