+++ webapp/src/pages/Trips.tsx
@@ -4,10 +4,10 @@ import LayoutContainer from "../components/LayoutContainer";
 import PageLoader from "../components/PageLoader";
 import { t } from "../localization";
 import { vehicleIconForVendorService } from "../modules/mobilityIconLookup";
-import Money from "../shared/react/Money";
 import useAsyncFetch from "../state/useAsyncFetch";
 import useUrlMarshal from "../state/useUrlMarshal";
 import Stack from "../ui/Stack";
+import Money from "../uir/Money";
 import dayjs from "dayjs";
 import isEmpty from "lodash/isEmpty";
 import React from "react";
