+++ webapp/src/pages/UnclaimedOrderList.tsx
@@ -10,7 +10,6 @@ import PageLoader from "../components/PageLoader";
 import SumaImage from "../components/SumaImage";
 import { t } from "../localization";
 import { dayjs } from "../modules/dayConfig";
-import ScrollTopOnMount from "../shared/ScrollToTopOnMount";
 import useAsyncFetch from "../state/useAsyncFetch";
 import Badge from "../ui/Badge";
 import Button from "../ui/Button";
@@ -19,6 +18,7 @@ import CardBody from "../ui/CardBody";
 import { Dialog } from "../ui/Dialog";
 import DialogHeader from "../ui/DialogHeader";
 import Stack from "../ui/Stack";
+import ScrollTopOnMount from "../uir/ScrollToTopOnMount";
 import isEmpty from "lodash/isEmpty";
 import React from "react";
