--- webapp/src/shared/README.md
@@ -1,8 +1,0 @@
-## DO NOT EDIT THESE FILES OUTSIDE OF `webapp`
-
-These files are shared between our JS/React apps,
-with the original source in `webapp/src/shared`.
-
-Use `make sync-shared` in the other app folders to pull in these files.
-In the future we can muse more sophisticated sharing mechanisms,
-but duplicating is fine for now.
